Collapse ShopiFixer flow to pricing and checkout
